Require login

+++ galleryproj/gallery/views.py
@@ -1,18 +1,22 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from django.urls import reverse
+from django.contrib.auth.decorators import login_required, permission_required
 
 from . import s3lib
 from .models import Photo, Like
 
+@login_required
 def home(request):
-    photos = Photo.objects.all().order_by('-upload_date')
+    photos = Photo.objects.all().filter(approved=True).order_by('-upload_date')
     for p in photos:
         p.liked = False
         for l in p.like_set.all():
             if request.user == l.user:
                 p.liked = True
                 break
+    print(request.user.has_perm('gallery.approve_photo'))
+    print(request.user.get_group_permissions())
     return render(request, 'home.html', {'photos':photos})
 
 def sign_s3(request):
@@ -21,6 +25,7 @@ def sign_s3(request):
     filetype = request.GET['file-type']
     return HttpResponse(s3lib.sign(filename,filetype))
 
+@login_required
 def submit_form(request):
     label = request.POST["label"]
     s3url = request.POST["img-url"]
@@ -29,20 +34,35 @@ def submit_form(request):
 
     return redirect('home')#, args=(question.id,)))
 
+@login_required
 def upload(request):
     photos = Photo.objects.all().filter(owner=request.user).order_by('-upload_date')
     return render(request, 'upload.html', {'photos':photos})
 
 def like(request):
-    l = Like(photo=Photo.objects.get(id=request.GET['photo-id']),
-        user=request.user
-    )
-    l.save()
+    if request.user.is_authenticated:
+        l = Like(photo=Photo.objects.get(id=request.GET['photo-id']),
+            user=request.user
+        )
+        l.save()
     return redirect('home')
 
 def dislike(request):
-    l = Like.objects.get(photo=Photo.objects.get(id=request.GET['photo-id']),
-        user=request.user
-    )
-    l.delete()
+    if request.user.is_authenticated:
+        l = Like.objects.get(photo=Photo.objects.get(id=request.GET['photo-id']),
+            user=request.user
+        )
+        l.delete()
     return redirect('home')
+
+@permission_required('gallery.can_approve')
+def approve(request):
+    if request.method == 'GET':
+        photos = Photo.objects.all().filter(approved=False)
+        return render(request, 'approve.html', {'photos': photos})
+    else:
+        p = Photo.objects.get(id=request.POST['photo-id'])
+        p.approved = True
+        p.save()
+        return redirect('approve')
+
